pyproject.toml
[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "xgboost-subsampling"
version = "0.1.0"
description = "Federated XGBoost with subsampling"
authors = ["The Flower Authors <hello@flower.ai>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
flwr-nightly = { extras = ["simulation"], version = ">=1.7.0,<2.0" }
flwr-datasets = ">=0.0.2,<1.0.0"
xgboost = ">=2.0.0,<3.0.0"
matplotlib = ">=3.8.4,<4.0.0"
